+++ 9.6/EM Food/Eco.EM.Food.Cafe/Items/Cheese.cs
@@ -32,7 +32,7 @@ namespace Eco.EM.Food.Cafe
 
         public override float Calories => EMFoodItemResolver.Obj.ResolveCalories(this);
         public override Nutrients Nutrition => EMFoodItemResolver.Obj.ResolveNutrients(this);
-        protected override int BaseShelfLife => EMFoodItemResolver.Obj.ResolveShelfLife(this);
+        protected override float BaseShelfLife => EMFoodItemResolver.Obj.ResolveShelfLife(this);
         static CheeseItem() => EMFoodItemResolver.AddDefaults(defaults);
     }
