+++ 9.6/EM Food/Eco.EM.Food.FoodSmoking/Items/PepperoniPizza.cs
@@ -33,7 +33,7 @@ namespace Eco.PlanetChefMod
         public override float Calories => 300;
         public override Nutrients Nutrition => new Nutrients() { Carbs = 33, Fat = 10, Protein = 15, Vitamins = 0 };
 
-        protected override int BaseShelfLife => throw new NotImplementedException();
+        protected override float BaseShelfLife => throw new NotImplementedException();
     }
 
     [RequiresSkill(typeof(CookingSkill), 1)]
