+++ 9.6/EM Food/Eco.EM.Food.Zymology/Items/SpecialtyGrains.cs
@@ -28,7 +28,7 @@ namespace Eco.EM.Food.Zymology
         public override Nutrients Nutrition => n;
         private Nutrients n => EMFoodItemResolver.Obj.ResolveNutrients(this);
 
-        protected override int BaseShelfLife => throw new NotImplementedException();
+        protected override float BaseShelfLife => throw new NotImplementedException();
 
         static SpecialtyGrainsItem() => EMFoodItemResolver.AddDefaults(defaults);
     }
